# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SIP Platform Control-Plane API
  version: "1.0.0"
  description: |
    REST API (`/v1`) for the multi-tenant SIP / UCaaS PBX control plane
    (Go control-plane). This documents the chi router in
    `control-plane/internal/api/server.go` and its handlers.

    ## Authentication

    Every `/v1` endpoint requires a Bearer token unless explicitly marked
    public:

    ```
    Authorization: Bearer sip_<token>
    ```

    Tokens look like `sip_<48hex>` and are either **tenant-scoped** or
    platform-wide (super-admin, `tenant_id` = null). A tenant admin can
    self-issue tenant-scoped keys in the portal, choosing a scope:

    - `read`  — GET endpoints only
    - `write` — read + create/update/delete
    - `admin` — full control (token / user / invite management)

    `401` = missing/invalid/expired token. `403` = token lacks the required
    scope or addresses a different tenant.

    ### Scope per endpoint

    Each operation's description states the **Required scope**. Routes wrapped
    with `RequireScope("admin"|"write"|...)` enforce that exact minimum.
    Routes without an explicit `RequireScope` wrapper still require a valid
    bearer token (any scope) but are not scope-gated beyond authentication —
    these are noted as **Required scope: authenticated**.

    ### Public endpoints (no auth)

    These bypass the auth middleware (`security: []`):

    - `GET  /v1/version`
    - `POST /v1/signup`
    - `POST /v1/invites/accept`
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/api-tokens:
    post:
      tags: [ApiTokens]
      summary: Create an API token
      description: |
        Mints a new bearer token (`sip_<48hex>`). The plaintext token is
        returned exactly once in the `token` field. Required scope: **admin**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPITokenRequest"
      responses:
        "201":
          description: Token created. Plaintext returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedAPIToken"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [ApiTokens]
      summary: List API tokens
      description: "Lists tokens (no plaintext, no hash). Required scope: **admin**."
      responses:
        "200":
          description: Tokens.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/api-tokens/{tokenID}:
    delete:
      tags: [ApiTokens]
      summary: Revoke an API token
      description: "Revokes a token by id. Required scope: **admin**."
      parameters:
        - $ref: "#/components/parameters/TokenID"
      responses:
        "204":
          description: Revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
  /v1/users:
    post:
      tags: [Users]
      summary: Create a portal user
      description: |
        Creates a portal login user. Only super-admin tokens may create
        `super_admin` users. Required scope: **admin**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/users/{userID}/password:
    post:
      tags: [Users]
      summary: Set a user's password
      description: "Overwrites a user's password (bcrypt server-side). Required scope: **admin**."
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
      responses:
        "204":
          description: Password set.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---------------------------------------------------------------------------
  # Public auth flows
  # ---------------------------------------------------------------------------
  /v1/signup:
    post:
      tags: [Auth]
      summary: Self-serve tenant signup
      description: |
        Public. Creates a fresh tenant and its first admin user in one
        transaction (the user is granted `tenant_admin` membership).
        Required scope: **public (no auth)**.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Tenant + user created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    $ref: "#/components/schemas/Tenant"
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/invites/accept:
    post:
      tags: [Auth]
      summary: Accept an invitation
      description: |
        Public. Consumes an invite token and creates/links the user account.
        Required scope: **public (no auth)**.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteRequest"
      responses:
        "200":
          description: Invite accepted; user returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid invite token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Invite already used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Invite expired or revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a password reset
      description: |
        Public. Always returns 202 (never reveals whether the email exists).
        Reset email is sent in the background. Required scope: **public (no auth)**.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Accepted (opaque).
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm a password reset
      description: |
        Public. Consumes a reset token and sets a new password.
        Required scope: **public (no auth)**.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  format: password
      responses:
        "200":
          description: Password reset; user returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Token already used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Token expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Invites (tenant-scoped admin)
  # ---------------------------------------------------------------------------
  /v1/tenants/{tenantID}/invites:
    post:
      tags: [Invites]
      summary: Create an invitation
      description: |
        Invites an email to join the tenant. The plaintext token is returned
        once (and emailed if SMTP is configured). Required scope: **admin**.
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedInvite"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Invites]
      summary: List invitations
      description: "Lists a tenant's invites (no plaintext token). Required scope: **admin**."
      parameters:
        - $ref: "#/components/parameters/TenantID"
      responses:
        "200":
          description: Invites.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/invites/{inviteID}:
    delete:
      tags: [Invites]
      summary: Revoke an invitation
      description: "Revokes a pending invite. Required scope: **admin**."
      parameters:
        - $ref: "#/components/parameters/InviteID"
      responses:
        "204":
          description: Revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ---------------------------------------------------------------------------
  # Tenants + SIP domains
  # ---------------------------------------------------------------------------
  /v1/tenants:
    post:
      tags: [Tenants]
      summary: Create a tenant
      description: "Creates a tenant. Required scope: **authenticated** (no explicit scope wrapper)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: Tenant created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Tenants]
      summary: List tenants
      description: "Lists tenants the caller can see. Required scope: **authenticated**."
      responses:
        "200":
          description: Tenants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/tenants/{tenantID}:
    get:
      tags: [Tenants]
      summary: Get a tenant
      description: "Fetches one tenant by id. Required scope: **authenticated**."
      parameters:
        - $ref: "#/components/parameters/TenantID"
      responses:
        "200":
          description: Tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/extensions/{extensionID}/features:
    patch:
      tags: [Extensions]
      summary: Update extension feature flags
      description: |
        PATCH-style update of feature flags. Only supplied fields change
        (nullable booleans distinguish "set false" from "unchanged").
        Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/ExtensionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExtensionFeaturesRequest"
      responses:
        "200":
          description: Updated extension.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # CDRs
  # ---------------------------------------------------------------------------
...
  /v1/devices/{mac}:
    get:
      tags: [Devices]
      summary: Get a device
      description: |
        Fetches one device by MAC. Unlike the list endpoint, this includes the
        `provisioning_token`. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/MAC"
      responses:
        "200":
          description: Device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/devices/{mac}/lines:
    post:
      tags: [Devices]
      summary: Bind a line to a device
      description: |
        Binds a line key (>=1) to an extension on the device. The extension
        must belong to the device's tenant. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/MAC"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [line_number, extension_id]
              properties:
                line_number:
                  type: integer
                  minimum: 1
                extension_id:
                  type: string
                  format: uuid
                label:
                  type: string
      responses:
        "201":
          description: Line bound.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceLine"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # DIDs
  # ---------------------------------------------------------------------------
...
  /v1/ring-groups/{ringGroupID}/members:
    post:
      tags: [RingGroups]
      summary: Add a ring group member
      description: |
        Adds an extension to a ring group. The extension must belong to the
        group's tenant. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/RingGroupID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [extension_id]
              properties:
                extension_id:
                  type: string
                  format: uuid
                priority:
                  type: integer
                ring_delay_sec:
                  type: integer
      responses:
        "201":
          description: Member added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RingGroupMember"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Queues
  # ---------------------------------------------------------------------------
  /v1/tenants/{tenantID}/queues:
    post:
      tags: [Queues]
      summary: Create a queue
      description: |
        Creates a mod_callcenter queue and (best-effort) syncs it to FreeSWITCH
        via ESL. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQueueRequest"
      responses:
        "201":
          description: Queue created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Queues]
      summary: List queues
      description: "Lists a tenant's call queues. Required scope: **authenticated**."
      parameters:
        - $ref: "#/components/parameters/TenantID"
      responses:
        "200":
          description: Queues.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Queue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/queues/{queueID}/agents:
    post:
      tags: [Queues]
      summary: Add a queue agent
      description: |
        Adds (or updates) an agent + tier for an extension and (best-effort)
        syncs to mod_callcenter. The extension must belong to the queue's
        tenant. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/QueueID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [extension_id]
              properties:
                extension_id:
                  type: string
                  format: uuid
                tier_level:
                  type: integer
                tier_position:
                  type: integer
                wrap_up_time:
                  type: integer
      responses:
        "201":
          description: Agent added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueAgent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # IVRs
  # ---------------------------------------------------------------------------
  /v1/tenants/{tenantID}/ivrs:
    post:
      tags: [IVRs]
      summary: Create an IVR
      description: "Creates an auto-attendant menu. Required scope: **authenticated**."
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIVRRequest"
      responses:
        "201":
          description: IVR created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IVR"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/ivrs/{ivrID}/options:
    post:
      tags: [IVRs]
      summary: Add an IVR option
      description: |
        Adds a digit option to a menu. For `action_kind=dial_e164`,
        `action_data` is normalized server-side. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/IVRID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [digit, action_kind]
              properties:
                digit:
                  type: string
                label:
                  type: string
                action_kind:
                  type: string
                action_id:
                  type: string
                  format: uuid
                  nullable: true
                action_data:
                  type: string
      responses:
        "201":
          description: Option added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IVROption"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Paging groups
  # ---------------------------------------------------------------------------
...
  /v1/carriers:
    get:
      tags: [Carriers]
      summary: List carriers
      description: "Lists available PSTN carriers. Required scope: **authenticated**."
      responses:
        "200":
          description: Carriers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Carrier"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/carriers/{kind}/accounts:
    post:
      tags: [Carriers]
      summary: Create a carrier account
      description: |
        Creates a per-tenant trunk (carrier account) under the carrier
        identified by `kind` (e.g. `callcentric`). `register` defaults to true.
        Required scope: **authenticated**.
      parameters:
        - name: kind
          in: path
          required: true
          description: Carrier kind (e.g. callcentric, telnyx).
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarrierAccountRequest"
      responses:
        "201":
          description: Carrier account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Carrier kind not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Voicemail
  # ---------------------------------------------------------------------------
  /v1/extensions/{extensionID}/voicemail:
    post:
      tags: [Voicemail]
      summary: Create a voicemail box
      description: |
        Creates a voicemail box for an extension and flips the extension's
        `voicemail_enabled` flag. `pin` is required. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/ExtensionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin:
                  type: string
                  description: 4-12 digits.
                email:
                  type: string
                  format: email
                timezone:
                  type: string
                max_messages:
                  type: integer
                max_msg_duration_sec:
                  type: integer
      responses:
        "201":
          description: Voicemail box created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoicemailBox"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Extension not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Voicemail]
      summary: Get a voicemail box
      description: "Fetches an extension's voicemail box. Required scope: **authenticated**."
      parameters:
        - $ref: "#/components/parameters/ExtensionID"
      responses:
        "200":
          description: Voicemail box.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoicemailBox"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No voicemail box for extension.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/extensions/{extensionID}/voicemail/messages:
    get:
      tags: [Voicemail]
      summary: List voicemail messages
      description: |
        Lists message metadata for an extension's box. The audio path is never
        serialized. Required scope: **authenticated**.
      parameters:
        - $ref: "#/components/parameters/ExtensionID"
      responses:
        "200":
          description: Messages (metadata only).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VoicemailMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No voicemail box for extension.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sip_<token>
      description: |
        Bearer token. Tokens look like `sip_<token>` (e.g. `sip_<48hex>`).
        Send as `Authorization: Bearer sip_<token>`.
...
  responses:
    BadRequest:
      description: Invalid request (bad JSON, missing/invalid fields).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    IssuedAPIToken:
      allOf:
        - $ref: "#/components/schemas/APIToken"
        - type: object
          properties:
            token:
              type: string
              description: Plaintext token (sip_<48hex>), shown once on create.
...